update: document code & small fixes

+++ ulysses_senteval/cache.py
@@ -1,4 +1,4 @@
-"""TODO."""
+"""Embed cache functions."""
 import typing as t
 import os
 
@@ -9,7 +9,7 @@ from . import utils
 
 
 def load_from_cache(cache_dir: str, task: str) -> t.Optional[torch.Tensor]:
-    """TODO."""
+    """Load cached embedding or return None if not found."""
     cache_output_uri = os.path.join(cache_dir, f"{task}.pt")
 
     if not os.path.exists(cache_output_uri):
@@ -18,8 +18,8 @@ def load_from_cache(cache_dir: str, task: str) -> t.Optional[torch.Tensor]:
     return torch.load(cache_output_uri).float()
 
 
-def save_in_cache(embs: utils.DataType, cache_dir: str, task: str) -> None:
-    """TODO."""
+def save_in_cache(embs: utils.DataType, cache_dir: str, task: str, overwrite: bool = False) -> None:
+    """Cache embedding."""
     if isinstance(embs, np.ndarray):
         embs = torch.from_numpy(embs)
 
@@ -27,4 +27,8 @@ def save_in_cache(embs: utils.DataType, cache_dir: str, task: str) -> None:
 
     os.makedirs(cache_dir, exist_ok=True)
     cache_output_uri = os.path.join(cache_dir, f"{task}.pt")
+
+    if not overwrite and os.path.exists(cache_output_uri):
+        raise RuntimeError(f"Embedding '{cache_output_uri}' exists but overwrite=False.")
+
     torch.save(embs, cache_output_uri)
